+++ app/src/main/java/com/absut/tasksapp/data/TaskRepositoryImpl.kt
@@ -17,13 +17,13 @@ class TaskRepositoryImpl @Inject constructor(private val taskDao: TaskDao) : Tas
 
     override suspend fun deleteTask(task: Task) = taskDao.delete(task)
 
-    override suspend fun deleteTaskById(id: Int) = taskDao.deleteById(id)
+    override suspend fun deleteTaskById(id: Long) = taskDao.deleteById(id)
 
     override suspend fun deleteCompletedTasks() = taskDao.deleteCompletedTasks()
 
     override suspend fun insertTask(task: Task) : Long = taskDao.insert(task)
 
-    override suspend fun getTaskById(id: Int) : Task = taskDao.getTaskById(id)
+    override suspend fun getTaskById(id: Long) : Task = taskDao.getTaskById(id)
 
-    override suspend fun updateTaskAsCompleted(id: Int) = taskDao.updateTaskAsCompleted(id)
+    override suspend fun updateTaskAsCompleted(id: Long) = taskDao.updateTaskAsCompleted(id)
 }
